move setup coursier + add tests
