+++ packages/metals-languageclient/src/__tests__/fetchMetals.test.ts
@@ -1,5 +1,4 @@
 import { calcServerDependency } from "../fetchMetals";
-import { validateCoursier } from "../fetchMetals";
 
 describe("fetchMetals", () => {
   describe("calcServerDependency", () => {
@@ -34,18 +33,5 @@ describe("fetchMetals", () => {
       const customVersion = "com.acme:metals_1.2.3:3.2.1-foobar";
       expect(calcServerDependency(customVersion)).toBe(customVersion);
     });
-  });
-
-  it("should find coursier in PATH", async () => {
-    const pathEnv = process.env["PATH"];
-    if (pathEnv) {
-      expect(await validateCoursier(pathEnv)).toBeDefined();
-    } else {
-      fail("PATH environment variable is not defined");
-    }
-  });
-
-  it("should not find coursier if not present in PATH", async () => {
-    expect(await validateCoursier("path/fake")).toBeUndefined();
   });
 });
